+++ build.sh
@@ -1,9 +1,7 @@
 #!/bin/bash
 
 cd "${0%/*}"
-plasmapkg2 -r org.kde.simpleMonitor
-# rm -rf ~/.local/share/plasma/plasmoids/org.kde.simpleMonitor
-plasmapkg2 -t plasmoid -i ./plasmoid
-kbuildsycoca5 --noincremental
-# plasmawindowed org.kde.simpleMonitor
-plasmoidviewer -a org.kde.simpleMonitor
+kpackagetool6 -t Plasma/Applet -r org.kde.simpleMonitor 2>/dev/null
+kpackagetool6 -t Plasma/Applet -i ./plasmoid
+kbuildsycoca6 --noincremental
+plasmoidviewer6 -a org.kde.simpleMonitor