+++ plasmoid/contents/ui/monitorWidgets/CpuWidget.qml
@@ -17,7 +17,7 @@
  * along with plasma-simpleMonitor.  If not, see <http://www.gnu.org/licenses/>.
  **/
 
-import QtQuick 2.0
+import QtQuick
 
 ListView {
     id: cpuListView
@@ -28,8 +28,7 @@ ListView {
 
     property color progressColor: "#993de515"
     
-    implicitWidth: 100 * units.devicePixelRatio
-    implicitHeight: childrenRect.height
+    implicitWidth: 100    implicitHeight: childrenRect.height
 
     model: cpuModel
     interactive: false
@@ -46,15 +45,13 @@ ListView {
     delegate: Item {
         id: itemElement
         width: cpuListView.width
-        height: cpuListItem.height + units.devicePixelRatio
+        height: cpuListItem.height + 1
         Column {
             id: cpuListItem
             width: parent.width
-            height: (20 + indicatorHeight) * units.devicePixelRatio
-            Row {
+            height: (20 + indicatorHeight)            Row {
                 spacing: 0
-                spacing: 5 * units.devicePixelRatio
-                anchors.left: parent.left
+                spacing: 5                anchors.left: parent.left
                 Text {
                     id: cpuLabel
                     text: i18n('CPU %1:', model.index)
@@ -77,8 +74,7 @@ ListView {
             }
             Item {
                 id: progressBar
-                height: 10 * units.devicePixelRatio
-                //clip: true
+                height: 10                //clip: true
                 width: parent.width
                 Rectangle {
                     // Clear background.
@@ -167,10 +163,7 @@ ListView {
                     }
                 }
                 Rectangle {
-                    height: progressBar.height + 4 * units.devicePixelRatio
-                    width: 5 * units.devicePixelRatio
-                    radius: 2 * units.devicePixelRatio
-                    anchors.left: rectValue.right
+                    height: progressBar.height + 4                    width: 5                    radius: 2                    anchors.left: rectValue.right
                     anchors.verticalCenter: parent.verticalCenter
                     color: "#88ffffff"
                     visible: !flatCpuLoad
@@ -191,7 +184,7 @@ ListView {
 
         ListView.onAdd: SequentialAnimation {
             PropertyAction { target: cpuListItem; property: "height"; value: 0 }
-            NumberAnimation { target: cpuListItem; property: "height"; to: 30 * units.devicePixelRatio; duration: 250; easing.type: Easing.InOutQuad }
+            NumberAnimation { target: cpuListItem; property: "height"; to: 30; duration: 250; easing.type: Easing.InOutQuad }
         }
     }