+++ plasmoid/contents/ui/monitorWidgets/GpuTempList.qml
@@ -17,7 +17,7 @@
  * along with plasma-simpleMonitor.  If not, see <http://www.gnu.org/licenses/>.
  **/
 
-import QtQuick 2.0
+import QtQuick
 
 ListView {
     id: gpuTempList
@@ -26,9 +26,7 @@ ListView {
     LayoutMirroring.enabled: direction === Qt.RightToLeft
     LayoutMirroring.childrenInherit: true
 
-    implicitHeight: count * 25 * units.devicePixelRatio
-    implicitWidth: 100 * units.devicePixelRatio
-
+    implicitHeight: count * 25    implicitWidth: 100
     interactive: false
 
     property int highTemp: 80
@@ -41,11 +39,9 @@ ListView {
 
     delegate: Item {
         id: gpuListTemp
-        implicitHeight: 25 * units.devicePixelRatio
-        implicitWidth: gpuLabel.implicitWidth + unitLabel.implicitWidth
+        implicitHeight: 25        implicitWidth: gpuLabel.implicitWidth + unitLabel.implicitWidth
         width: parent.width
-        height: (20 + indicatorHeight) * units.devicePixelRatio
-        visible: showGpuTemp
+        height: (20 + indicatorHeight)        visible: showGpuTemp
         Text {
             id: gpuLabel
             anchors.left: parent.left
@@ -66,18 +62,17 @@ ListView {
 
         Rectangle {
             id: rectValue
-            height: indicatorHeight * units.devicePixelRatio
-            width: Math.round(val/gpuTempList.maxTemp*parent.width)
+            height: indicatorHeight            width: Math.round(val/gpuTempList.maxTemp*parent.width)
             color: if (Math.round(val) >= coreTempList.criticalTemp) "red"
                    else if (Math.round(val) >= gpuTempList.highTemp) "#ffac2a"
                    else "#85a9ff"
             anchors.top: gpuLabel.bottom
             anchors.right: parent.right
-            anchors.topMargin: units.devicePixelRatio
+            anchors.topMargin: 1
         }
         ListView.onAdd: SequentialAnimation {
             PropertyAction { target: gpuListTemp; property: "height"; value: 0 }
-            NumberAnimation { target: gpuListTemp; property: "height"; to: (20 + indicatorHeight) * units.devicePixelRatio; duration: 250; easing.type: Easing.InOutQuad }
+            NumberAnimation { target: gpuListTemp; property: "height"; to: (20 + indicatorHeight); duration: 250; easing.type: Easing.InOutQuad }
         }
     }