+++ plasmoid/contents/ui/monitorWidgets/TimePicker.qml
@@ -17,7 +17,7 @@
  * along with plasma-simpleMonitor.  If not, see <http://www.gnu.org/licenses/>.
  **/
 
-import QtQuick 2.0
+import QtQuick
 
 Item {
     id: timePicker
@@ -26,8 +26,7 @@ Item {
     LayoutMirroring.childrenInherit: true
 
     implicitWidth: secs.x + secs.width
-    implicitHeight: hour.implicitHeight - 11 * units.devicePixelRatio
-
+    implicitHeight: hour.implicitHeight - 11
     FontLoader {
         id: playRegular
         source: "../../fonts/Play/Play-Regular.ttf"
@@ -50,8 +49,7 @@ Item {
         id: hour
         anchors.left: parent.left
         anchors.bottom: parent.bottom
-        height: 44 * units.devicePixelRatio
-        text: "00:"
+        height: 44        text: "00:"
         font {
             family: playRegular.name
             pointSize: 30
@@ -62,8 +60,7 @@ Item {
         id: mins
         anchors.left: hour.right
         anchors.bottom: parent.bottom
-        height: 29 * units.devicePixelRatio
-        text: "00:"
+        height: 29        text: "00:"
         font {
             family: playRegular.name
             pointSize: 20
@@ -74,8 +71,7 @@ Item {
         id: secs
         anchors.left: mins.right
         anchors.bottom: parent.bottom
-        height: 21 * units.devicePixelRatio
-        text: "00"
+        height: 21        text: "00"
         font {
             family: playRegular.name
             pointSize: 15