+++ plasmoid/contents/ui/skins/ColumnSkin.qml
@@ -17,9 +17,9 @@
  * along with plasma-simpleMonitor.  If not, see <http://www.gnu.org/licenses/>.
  **/
 
-import QtQuick 2.0
-import QtQuick.Layouts 1.1
-import org.kde.plasma.components 2.0 as PlasmaComponents
+import QtQuick
+import QtQuick.Layouts
+import QtQuick.Controls as Controls
 
 import "../monitorWidgets"
 import "../components"
@@ -39,8 +39,7 @@ BaseSkin {
         id: mainLayout
 
         anchors.fill: parent
-        anchors.margins: 5 * units.devicePixelRatio
-        spacing: 0
+        anchors.margins: 5        spacing: 0
 
         Item {
             implicitHeight: datePicker.implicitHeight
@@ -72,8 +71,7 @@ BaseSkin {
             rowSpacing: 0
 
             Layout.fillWidth: true
-            Layout.topMargin: 5 * units.devicePixelRatio
-
+            Layout.topMargin: 5
             ColumnLayout {
                 id: distroInfo
 
@@ -87,9 +85,7 @@ BaseSkin {
 
                 LogoImage {
                     id: distroLogo
-                    Layout.minimumWidth: (implicitWidth < implicitHeight) ? 100*implicitWidth/implicitHeight : 100 * units.devicePixelRatio
-                    Layout.minimumHeight: (implicitHeight < implicitWidth) ? 100*implicitHeight/implicitWidth : 100 * units.devicePixelRatio
-                    Layout.preferredWidth: (Layout.fillWidth) ? Layout.minimumWidth : height * implicitWidth/implicitHeight
+                    Layout.minimumWidth: (implicitWidth < implicitHeight) ? 100*implicitWidth/implicitHeight : 100                    Layout.minimumHeight: (implicitHeight < implicitWidth) ? 100*implicitHeight/implicitWidth : 100                    Layout.preferredWidth: (Layout.fillWidth) ? Layout.minimumWidth : height * implicitWidth/implicitHeight
                     Layout.preferredHeight: (Layout.fillHeight) ? Layout.minimumHeight : width * implicitHeight/implicitWidth
                     Layout.fillWidth: (implicitWidth < implicitHeight) ? false: true
                     Layout.fillHeight: !Layout.fillWidth
@@ -106,12 +102,12 @@ BaseSkin {
                         onClicked: logoPopup.open(mouse.x, mouse.y)
                     }
 
-                    PlasmaComponents.ContextMenu {
+                    Controls.Menu {
                         id: logoPopup
 
-                        PlasmaComponents.MenuItem {
+                        Controls.MenuItem {
                             text: distroLogo.editMode ? i18n("Lock image scaling") : i18n("Unlock image scaling")
-                            onClicked: distroLogo.editMode = !distroLogo.editMode
+                            onTriggered: distroLogo.editMode = !distroLogo.editMode
                         }
                     }
                 }
@@ -126,8 +122,7 @@ BaseSkin {
                     kernelVersion: root.kernelVersion
 
                     Layout.alignment: Qt.AlignHCenter | Qt.AlignTop
-                    Layout.topMargin: 2 * units.devicePixelRatio
-                    Layout.minimumHeight: implicitHeight
+                    Layout.topMargin: 2                    Layout.minimumHeight: implicitHeight
                     Layout.maximumHeight: implicitHeight
                     Layout.minimumWidth: implicitWidth
                     Layout.preferredWidth: implicitWidth
@@ -139,10 +134,7 @@ BaseSkin {
                 id: timePicker
 
                 Layout.alignment: Qt.AlignLeft
-                Layout.topMargin: -5 * units.devicePixelRatio
-                Layout.leftMargin: 10 * units.devicePixelRatio
-                Layout.bottomMargin: 5 * units.devicePixelRatio
-                Layout.minimumHeight: implicitHeight
+                Layout.topMargin: -5                Layout.leftMargin: 10                Layout.bottomMargin: 5                Layout.minimumHeight: implicitHeight
                 Layout.maximumHeight: implicitHeight
                 Layout.preferredWidth: implicitWidth
                 Layout.preferredHeight: implicitHeight
@@ -153,10 +145,7 @@ BaseSkin {
 
                 Layout.leftMargin: 2
                 Layout.fillWidth: true
-                Layout.minimumHeight: 3 * units.devicePixelRatio
-                Layout.maximumHeight: 3 * units.devicePixelRatio
-                Layout.preferredHeight: 3 * units.devicePixelRatio
-            }
+                Layout.minimumHeight: 3                Layout.maximumHeight: 3                Layout.preferredHeight: 3            }
 
 
             CoreTempList {
@@ -168,10 +157,7 @@ BaseSkin {
                 tempUnit: root.tempUnit
                 direction: root.direction
 
-                Layout.leftMargin: 5 * units.devicePixelRatio
-                Layout.rightMargin: 5 * units.devicePixelRatio
-                Layout.topMargin: 5 * units.devicePixelRatio
-                Layout.fillWidth: true
+                Layout.leftMargin: 5                Layout.rightMargin: 5                Layout.topMargin: 5                Layout.fillWidth: true
                 Layout.fillHeight: true
                 Layout.minimumWidth: implicitWidth
                 Layout.minimumHeight: implicitHeight
@@ -187,10 +173,7 @@ BaseSkin {
                     tempUnit: root.tempUnit
                     direction: root.direction
     
-                    Layout.leftMargin: 5 * units.devicePixelRatio
-                    Layout.rightMargin: 5 * units.devicePixelRatio
-                    Layout.topMargin: 5 * units.devicePixelRatio
-                    Layout.fillWidth: true
+                    Layout.leftMargin: 5                    Layout.rightMargin: 5                    Layout.topMargin: 5                    Layout.fillWidth: true
                     Layout.fillHeight: false
                     Layout.minimumWidth: implicitWidth
                     Layout.minimumHeight: implicitHeight
@@ -207,10 +190,7 @@ BaseSkin {
                     tempUnit: root.tempUnit
                     direction: root.direction
     
-                    Layout.leftMargin: 5 * units.devicePixelRatio
-                    Layout.rightMargin: 5 * units.devicePixelRatio
-                    Layout.topMargin: 1 * units.devicePixelRatio
-                    Layout.fillWidth: true
+                    Layout.leftMargin: 5                    Layout.rightMargin: 5                    Layout.topMargin: 1                    Layout.fillWidth: true
                     Layout.fillHeight: true
                     Layout.minimumWidth: implicitWidth
                     Layout.minimumHeight: implicitHeight
@@ -224,20 +204,14 @@ BaseSkin {
             color: "white"
 
             Layout.fillWidth: true
-            Layout.minimumHeight: 3 * units.devicePixelRatio
-            Layout.maximumHeight: 3 * units.devicePixelRatio
-            Layout.preferredHeight: 3 * units.devicePixelRatio
-            Layout.topMargin: 5 * units.devicePixelRatio
-        }
+            Layout.minimumHeight: 3            Layout.maximumHeight: 3            Layout.preferredHeight: 3            Layout.topMargin: 5        }
 
         CpuWidget {
             id: cpuList
 
             direction: root.direction
 
-            Layout.leftMargin: 5 * units.devicePixelRatio
-            Layout.topMargin: 5 * units.devicePixelRatio
-            Layout.fillWidth: true
+            Layout.leftMargin: 5            Layout.topMargin: 5            Layout.fillWidth: true
             Layout.minimumWidth: implicitWidth
             Layout.minimumHeight: implicitHeight
             Layout.preferredWidth: implicitWidth
@@ -248,11 +222,7 @@ BaseSkin {
             color: "white"
 
             Layout.fillWidth: true
-            Layout.minimumHeight: 3 * units.devicePixelRatio
-            Layout.maximumHeight: 3 * units.devicePixelRatio
-            Layout.preferredHeight: 3 * units.devicePixelRatio
-            Layout.topMargin: 5 * units.devicePixelRatio
-        }
+            Layout.minimumHeight: 3            Layout.maximumHeight: 3            Layout.preferredHeight: 3            Layout.topMargin: 5        }
 
         MemArea {
             id: memArea
@@ -264,10 +234,7 @@ BaseSkin {
             memBuffers: root.memBuffers
 
             Layout.columnSpan: 2
-            Layout.topMargin: 2 * units.devicePixelRatio
-            Layout.leftMargin: 10 * units.devicePixelRatio
-            Layout.rightMargin: 5 * units.devicePixelRatio
-            Layout.fillWidth: true
+            Layout.topMargin: 2            Layout.leftMargin: 10            Layout.rightMargin: 5            Layout.fillWidth: true
             Layout.minimumWidth: implicitWidth
             Layout.minimumHeight: implicitHeight
             Layout.preferredWidth: implicitWidth
@@ -279,11 +246,7 @@ BaseSkin {
             color: "white"
 
             Layout.fillWidth: true
-            Layout.minimumHeight: 3 * units.devicePixelRatio
-            Layout.maximumHeight: 3 * units.devicePixelRatio
-            Layout.preferredHeight: 3 * units.devicePixelRatio
-            Layout.topMargin: 5 * units.devicePixelRatio
-        }
+            Layout.minimumHeight: 3            Layout.maximumHeight: 3            Layout.preferredHeight: 3            Layout.topMargin: 5        }
 
         MemArea {
             id: swapArea
@@ -295,10 +258,7 @@ BaseSkin {
             memUsed: root.swapUsed
 
             Layout.columnSpan: 2
-            Layout.topMargin: 2 * units.devicePixelRatio
-            Layout.leftMargin: 10 * units.devicePixelRatio
-            Layout.rightMargin: 5 * units.devicePixelRatio
-            Layout.fillWidth: true
+            Layout.topMargin: 2            Layout.leftMargin: 10            Layout.rightMargin: 5            Layout.fillWidth: true
             Layout.minimumWidth: implicitWidth
             Layout.minimumHeight: implicitHeight
             Layout.preferredWidth: implicitWidth